Refactor to generate typescript from server schema
Pull core logic into a graphql module and generate a
graphql schema SDL as part of the build.  Have the js
client depend on that SDL instead of hitting the deployed
service to get the SDL.  Pull out emails from js client
to avoid circular dependencies.  Finally, have the server
module pull in client and graphql modules to serve the
final application.